PowerModelsGMD.jl Library Reference

PowerModelsGMD.jl is a Julia package built on top of PowerModels.jl and InfrastructureModels.jl for analysis of geomagnetic disturbances in the context of power systems optimization problems.

Problem Specifications

GIC-Only

PowerModelsGMD.solve_gmd — Function

Solve GIC current model with an optimizer given input file in extended MatPower format

Solve GIC current model with Lehtinen–Pirjola (LP) matrix solver given RAW/GIC/CSV file path inputs

Solve GIC current model with Lehtinen–Pirjola (LP) matrix solver given RAW/GIC/CSV file handle inputs

Solve GIC current model with Lehtinen–Pirjola (LP) matrix solver given with RAW/GICfile paths and specified field magnitude/direction for a uniform electric field

Solve GIC current model with Lehtinen–Pirjola (LP) matrix solver given extended MatPower file path

Solve GIC current model with Lehtinen–Pirjola (LP) matrix solver given dictionary input

PowerModelsGMD.build_gmd — Function

Build the quasi-dc-pf problem as a linear constraint satisfaction problem

PowerModelsGMD.solve_gmd_ts_decoupled — Function

Solve the multi-time-series quasi-dc-pf problem

GIC-PF

PowerModelsGMD.solve_gmd_pf — Function

Solve coupled GMD PF problem

PowerModelsGMD.build_gmd_pf — Function

Build the coupled GMD PF problem

PowerModelsGMD.solve_ac_gmd_pf_uncoupled — Function

Solve sequential GMD PF problem with nonlinear ac equations

PowerModelsGMD.solve_soc_gmd_pf_uncoupled — Function

Solve sequential GMD PF problem with second order cone relaxation

PowerModelsGMD.solve_qc_gmd_pf_uncoupled — Function

Solve sequential GMD PF problem with nonlinear ac equations

GIC-OPF

PowerModelsGMD.solve_ac_gmd_opf — Function

Solve basic GMD OPF model with nonlinear ac polar

PowerModelsGMD.solve_soc_gmd_opf — Function

Solve basic GMD OPF model with second order cone ac polar relaxation

PowerModelsGMD.solve_gmd_opf — Function

Solve basic GMD OPF model

PowerModelsGMD.build_gmd_opf — Function

Build the coupled quasi-dc-pf and ac-opf problem as generator dispatch minimization problem

[1] M. Ryu, H. Nagarajan and R. Bent, Mitigating the Impacts of Uncertain Geomagnetic Disturbances on Electric Grids: A Distributionally Robust Optimization Approach, in IEEE Transactions on Power Systems, vol. 37, no. 6, pp. 4258-4269, Nov. 2022, doi: 10.1109/TPWRS.2022.3147104.

*Note - [1] solves a robust variation of buildgmdopf and is used as a reference for the GIC equations

PowerModelsGMD.solve_ac_gmd_opf_uncoupled — Function

Solve basic GMD OPF model with nonlinear ac polar where the gic is an input parameter (uncoupled)

PowerModelsGMD.solve_soc_gmd_opf_uncoupled — Function

Solve basic GMD OPF model with second order cone ac polar relaxation

PowerModelsGMD.solve_gmd_opf_uncoupled — Function

Solve basic GMD OPF model

PowerModelsGMD.build_gmd_opf_uncoupled — Function

Build the coupled quasi-dc-pf and ac-opf problem as generator dispatch minimization problem

GIC-MLD

PowerModelsGMD.solve_ac_gmd_mld — Function

Solve GMD MLD mitigation with nonlinear ac equations

PowerModelsGMD.solve_soc_gmd_mld — Function

Solve GMD MLD mitigation with second order cone relaxation

PowerModelsGMD.solve_qc_gmd_mld — Function

Solve GMD MLD mitigation with nonlinear ac equations

PowerModelsGMD.build_gmd_mld — Function

Build the ac minimum loadshedding coupled with quasi-dc-pf problem as a maximum loadability problem with relaxed generator and bus participation

PowerModelsGMD.solve_ac_gmd_mld_uncoupled — Function

Solve GMD MLD mitigation with nonlinear ac equations

PowerModelsGMD.solve_soc_gmd_mld_uncoupled — Function

Solve GMD MLD mitigation with second order cone relaxation

PowerModelsGMD.solve_qc_gmd_mld_uncoupled — Function

Solve GMD MLD mitigation with nonlinear ac equations

PowerModelsGMD.build_gmd_mld_uncoupled — Function

Build the ac minimum loadshedding coupled with quasi-dc-pf problem as a maximum loadability problem on shunts and loads where the dc pf flow is uncoupled from the ac calculations

GIC-OTS

PowerModelsGMD.solve_ac_gmd_ots — Function

Solve GMD OTS mitigation with nonlinear ac polar relaxation

PowerModelsGMD.solve_soc_gmd_ots — Function

Solve GMD MLS OTS mitigation with second order cone relaxation

PowerModelsGMD.solve_qc_gmd_ots — Function

Solve GMD MLS OTS mitigation with quadratic constrained least squares relaxation

Blocker Placement

PowerModelsGMD.solve_ac_blocker_placement — Function

Run GMD mitigation with nonlinear ac equations

PowerModelsGMD.solve_soc_blocker_placement — Function

Run GMD mitigation with second order cone relaxation

PowerModelsGMD.build_blocker_placement — Function

Build the ac minimum loadshed coupled with quasi-dc power flow problem as a maximum loadability problem with relaxed generator and bus participation

PowerModelsGMD.solve_ac_blocker_placement_multi_scenario — Function

Run GMD mitigation with nonlinear ac equations

PowerModelsGMD.solve_soc_blocker_placement_multi_scenario — Function

Run GMD mitigation with second order cone relaxation

PowerModelsGMD.build_blocker_placement_multi_scenario — Function

Build the multi-scenario or time-series blocker placement problem

Objectives

PowerModelsGMD.objective_blocker_placement_cost — Function

Minimize cost of installing GIC blockers

PowerModelsGMD.objective_blocker_placement_cost_multi_scenario — Function

Minimize cost of installing GIC blockers for a multinetwork problem. Only the first network is considered when calculating cost.

PowerModelsGMD.objective_max_qloss — Function

Maximize the total qloss in the ac model

PowerModelsGMD.objective_bound_gmd_bus_v — Function

Max/min of the dc voltage at the substation

PowerModelsGMD.objective_max_loadability — Function

Maximize power delivered to loads

PowerModelsGMD.objective_gmd_min_fuel — Function

Minimize fuel

PowerModelsGMD.objective_gmd_min_error — Function

Minimize the sum-squared error between generator powers and their real/reactive power setpoints

PowerModelsGMD.objective_gmd_mls_on_off — Function

Minimize load shedding and fuel cost

PowerModelsGMD.objective_gmd_min_mls — Function

Minimize load shedding and fuel cost

PowerModelsGMD.objective_gmd_min_transformer_heating — Function

Minimize transfomer heating caused by GIC

Constraints

PowerModelsGMD.constraint_load_served — Function

CONSTRAINT: more than a specified percentage of load is served

CONSTRAINT: more than a specified percentage of load is served

PowerModelsGMD.constraint_gmd_connections — Function

CONSTRAINT: that ensures that the connection does not become ungrounded

Voltage Constraints

PowerModelsGMD.constraint_model_voltage — Function

Constraint: constraints on modeling bus voltages that is primarly a pass through to PM.constraintmodel_voltage There are a few situations where the GMD problem formulations have additional voltage modeling than what _PM provides. This adds connection between w and vm variables in the WR formulation space

Constraint: constraints on modeling bus voltages that is primarly a pass through to PM.constraintmodel_voltage There are a few situations where the GMD problem formulations have additional voltage modeling than what _PM provides. For example, many of the GMD problem formulations need explict vm variables, which the WR formulations do not provide

PowerModelsGMD.constraint_model_voltage_on_off — Function

Constraint: constraints on modeling bus voltages that is primarly a pass through to PMR.constraintbusvoltageon_off( There are a few situations where the GMD problem formulations have additional voltage modeling than what _PM provides. This adds connection between w and vm variables in the WR formulation space

Constraint: constraints on modeling bus voltages that is primarly a pass through to PMR.constraintbusvoltageon_off There are a few situations where the GMD problem formulations have additional voltage modeling than what _PMR provides. For example, many of the GMD problem formulations need explict vm variables, which the WR formulations do not provide

DC Current Constraints

PowerModelsGMD.constraint_dc_current_mag_gwye_delta_xf — Function

CONSTRAINT: dc current on ungrounded gwye-delta transformers

CONSTRAINT: dc current on ungrounded gwye-delta transformers

CONSTRAINT: dc current on ungrounded gwye-delta transformers

CONSTRAINT: dc current on ungrounded gwye-delta transformers

CONSTRAINT: dc current on ungrounded gwye-delta transformers

CONSTRAINT: dc current on ungrounded gwye-delta transformers

PowerModelsGMD.constraint_dc_current_mag_gwye_gwye_xf — Function

CONSTRAINT: dc current on ungrounded gwye-gwye transformers

CONSTRAINT: dc current on ungrounded gwye-gwye transformers

CONSTRAINT: dc current on ungrounded gwye-gwye transformers

CONSTRAINT: dc current on ungrounded gwye-gwye transformers

CONSTRAINT: dc current on ungrounded gwye-gwye transformers

CONSTRAINT: dc current on ungrounded gwye-gwye transformers

PowerModelsGMD.constraint_dc_current_mag_gwye_gwye_auto_xf — Function

CONSTRAINT: dc current on ungrounded gwye-gwye auto transformers

CONSTRAINT: dc current on ungrounded gwye-gwye auto transformers

CONSTRAINT: dc current on ungrounded gwye-gwye auto transformers

CONSTRAINT: dc current on ungrounded gwye-gwye auto transformers

CONSTRAINT: dc current on ungrounded gwye-gwye auto transformers

CONSTRAINT: dc current on ungrounded gwye-gwye auto transformers

Power Balance Constraints

PowerModelsGMD.constraint_power_balance_gmd — Function

CONSTRAINT: nodal power balance with gmd

CONSTRAINT: nodal power balance with gmd

CONSTRAINT: nodal power balance with gmd

CONSTRAINT: nodal power balance with gmd

CONSTRAINT: nodal power balance with gmd

PowerModelsGMD.constraint_power_balance_gmd_shunt — Function

CONSTRAINT: nodal power balance with gmd and shunts

CONSTRAINT: nodal power balance with gmd and shunts

CONSTRAINT: nodal power balance with gmd and shunts

CONSTRAINT: nodal power balance with gmd and shunts

CONSTRAINT: nodal power balance with gmd and shunts

PowerModelsGMD.constraint_power_balance_gmd_shunt_ls — Function

CONSTRAINT: nodal power balance with gmd, shunts, and constant power factor load shedding

CONSTRAINT: nodal power balance with gmd, shunts, and constant power factor load shedding

CONSTRAINT: nodal power balance with gmd, shunts, and constant power factor load shedding

CONSTRAINT: nodal power balance with gmd, shunts, and constant power factor load shedding

CONSTRAINT: nodal power balance with gmd, shunts, and constant power factor load shedding

PowerModelsGMD.constraint_dc_kcl — Function

CONSTRAINT: nodal kcl for dc circuits with shunts

CONSTRAINT: nodal kcl for dc circuits

PowerModelsGMD.constraint_dc_power_balance_ne_blocker — Function

CONSTRAINT: nodal power balance for dc circuits with GIC blockers and shunts

CONSTRAINT: nodal power balance for dc circuits with GIC blockers

PowerModelsGMD.constraint_dc_kcl_ne_blocker — Function

CONSTRAINT: nodal current balance for dc circuits with GIC blockers and shunts

CONSTRAINT: nodal power balance for dc circuits with GIC blockers

CONSTRAINT: nodal current balance for dc circuits with GIC blockers

Ohm's Law Constraints

PowerModelsGMD.constraint_dc_ohms — Function

CONSTRAINT: ohms constraint for dc circuits

CONSTRAINT: ohms constraint for dc circuits

Qloss Constraints

PowerModelsGMD.constraint_qloss_gmd — Function

CONSTRAINT: Calculation of qloss on a per edge basis gmd where vm is parameter

PowerModelsGMD.constraint_qloss — Function

CONSTRAINT: Calculation of qloss on a per edge basis

CONSTRAINT: qloss calculcated from ac voltage and dc current

CONSTRAINT: qloss assuming constant ac primary voltage

PowerModelsGMD.constraint_qloss_pu — Function

CONSTRAINT: Calculation of qloss on a per edge basis

CONSTRAINT: qloss calculcated for ac formulation single phase

CONSTRAINT: relaxed qloss calculcated for ac formulation single phase

PowerModelsGMD.constraint_qloss_constant_ieff — Function

CONSTRAINT: Calculation of qloss on a per edge basis where ieff is a constant

CONSTRAINT: qloss calculcated from ac voltage and constant ieff

Transformer Temperature Constraints

PowerModelsGMD.constraint_temperature_state — Function

CONSTRAINT: temperature state

CONSTRAINT: temperature state

CONSTRAINT: temperature state

PowerModelsGMD.constraint_temperature_state_ss — Function

CONSTRAINT: steady-state temperature state

PowerModelsGMD.constraint_hotspot_temperature_state_ss — Function

CONSTRAINT: steady-state hot-spot temperature state

PowerModelsGMD.constraint_hotspot_temperature_state — Function

CONSTRAINT: hot-spot temperature state

PowerModelsGMD.constraint_absolute_hotspot_temperature_state — Function

CONSTRAINT: absolute hot-spot temperature state

Variables

Not found: variable_iv

PowerModelsGMD.variable_bus_voltage — Function

Declaration of the bus voltage variables. This is a pass through to PM.variablebus_voltage except for those forms where vm is not created and it is needed for the GIC. For example, the WR models are formulated in the space of V^2, so there is no vm variable, so those variables need to be created

Declaration of the bus voltage variables. This is a pass through to PM.variablebus_voltage except for those forms where vm is not created and it is needed for the GIC. This function creates the vm variables to add to the WR formulation

PowerModelsGMD.variable_bus_voltage_on_off — Function

Declaration of the bus voltage variables. This is a pass through to PMR.variablebusvoltageon_off except for those forms where vm is not created and it is needed for the GIC. For example, the WR models are formulated in the space of V^2, so there is no vm variable, so those variables need to be created

Declaration of the bus voltage variables. This is a pass through to PM.variablebus_voltage except for those forms where vm is not created and it is needed for the GIC. This function creates the vm variables to add to the WR formulation

PowerModelsGMD.variable_gic_current — Function

VARIABLE: Declaration of variables associated with modeling of injected GIC

PowerModelsGMD.variable_dc_current_mag — Function

VARIABLE: dc current magnitude

PowerModelsGMD.variable_ac_positive_current_mag — Function

VARIABLE: ac current magnitude

PowerModelsGMD.variable_ac_current_mag_sqr — Function

VARIABLE: dc current magnitude squared

PowerModelsGMD.variable_ne_blocker_indicator — Function

VARIABLE: gic ne_blocker indicator

Voltage Variables

PowerModelsGMD.variable_dc_voltage — Function

VARIABLE: bus dc voltage

PowerModelsGMD.variable_dc_voltage_difference — Function

VARIABLE: bus dc voltage difference

PowerModelsGMD.variable_dc_voltage_on_off — Function

VARIABLE: bus dc voltage on/off

Power Balance Variables

Not found: variable_dc_bus_flow

PowerModelsGMD.variable_dc_line_flow — Function

VARIABLE: dc line flow

PowerModelsGMD.variable_dc_gen_flow — Function

VARIABLE: gen xfrm flow

Qloss Variables

PowerModelsGMD.variable_qloss — Function

VARIABLE: qloss

Transformer Heating Variables

PowerModelsGMD.variable_hotspot — Function

VARIABLE: hot-spot temperature

PowerModelsGMD.variable_delta_hotspot_ss — Function

VARIABLE: steady-state hot-spot temperature rise

PowerModelsGMD.variable_delta_hotspot — Function

VARIABLE: hot-spot temperature rise

PowerModelsGMD.variable_delta_oil_ss — Function

VARIABLE: steady-state top-oil temperature rise

PowerModelsGMD.variable_delta_oil — Function

VARIABLE: top-oil temperature rise

Solution Builders

PowerModelsGMD.solution_gmd_qloss! — Function

SOLUTION: add gmd qloss solution

Adds an explict gmd_qloss term for each branch

PowerModelsGMD.solution_gmd! — Function

SOLUTION: add quasi-dc power flow solutions

Adds an explict gmd_idc term for each branch ieff for all branches qloss for all branches

PowerModelsGMD.solution_gmd — Function

solution for matrix gms nl_solver
    returns dc bus voltages, neutral voltages and currents

PowerModelsGMD.add_ieff_solution! — Function

Add ieff field from solution into network

PowerModelsGMD.source_id_keys! — Function

Re-index solution values by component source_id

PowerModelsGMD.solution_add_qloss_bound_case! — Function

Set upper bound for branch qloss based on qloss maximization solution

PowerModelsGMD.solution_get_qloss_bound — Function

Get upper bound for branch qloss from qloss maximization solution

PowerModelsGMD.solution_add_gmd_bus_v_bounds_case! — Function

Add GMD bus voltage bounds into the case based on voltage maximization/minimization solutions

PowerModelsGMD.solution_get_gmd_bus_v_bounds — Function

Get GMD bus voltage bounds based on voltage maximization/minimization solutions

Reference Dictionary Builders

PowerModelsGMD.ref_add_gmd! — Function

REF: add gmd data to PowerModels.jl ref dict structures

PowerModelsGMD.ref_add_ne_blocker! — Function

REF: add expansion blocker data (ne_blocker) to PowerModels.jl ref dict structures

PowerModelsGMD.ref_add_ieff! — Function

REF: add ieff solution to branches

PowerModelsGMD.ref_add_gmd_connections! — Function

REF: adds connections to ref

PowerModelsGMD.ref_add_transformers! — Function

REF: adds transformers to ref

Base

PowerModelsGMD.check_gmd_branch_parent_status — Function

Check gmd branch parent status

Data Management

PowerModelsGMD.apply_mods! — Function

Apply mods specified by a path to a JSON file

Apply mods given a file stream to a JSON file

Apply a dictionary of mods

PowerModelsGMD.fix_gmd_indices! — Function

Correct parent branches for gmd branches after applying mods

PowerModelsGMD.create_sid_map — Function

Return a dictonary indexing network components by their source IDs

PowerModelsGMD.add_gmd_data! — Function

Add results from GMD solve into the ac case data structure

PowerModelsGMD.make_gmd_mixed_units! — Function

Make GMD mixed units

PowerModelsGMD.make_gmd_per_unit! — Function

Make GMD per unit

Make GMD per unit

PowerModelsGMD.add_connected_components! — Function

Add connected components to the network data structure. This information is used when setting variable bounds associated with GMD quantities

PowerModelsGMD.get_connected_components — Function

Return connected components for the network data structure. This information is used when setting variable bounds associated with GMD quantities

PowerModelsGMD.add_blockers! — Function

Add network expansion blocker objects at all substation neutral GMD buses

PowerModelsGMD.filter_gmd_ne_blockers! — Function

Remove network expansion blockers corresponding to connected components with a cardinality of 1 or less

PowerModelsGMD.update_cost_multiplier! — Function

Scale network expansion blocker cost proportional to the number of transfomers present at the corresponding substation

PowerModelsGMD.create_ts_net — Function

Create a multinetwork object from a single network and dictionary structure of time-series coupled voltage waveforms

Transformer Ieff Calculation

PowerModelsGMD.calc_ieff_current_mag — Function

Calculate ieff current magnitude for branches

PowerModelsGMD.calc_ieff_current_mag_line — Function

Calculate dc current on normal lines

PowerModelsGMD.calc_ieff_current_mag_grounded_xf — Function

Calculate dc current on grounded transformers

PowerModelsGMD.calc_ieff_current_mag_gwye_delta_xf — Function

Calculate dc current on ungrounded gwye-delta transformers

PowerModelsGMD.calc_ieff_current_mag_gwye_gwye_xf — Function

Calculate dc current on ungrounded gwye-gwye transformers

PowerModelsGMD.calc_ieff_current_mag_gwye_gwye_auto_xf — Function

Calculate dc current on ungrounded gwye-gwye auto transformers

PowerModelsGMD.calc_ieff_current_mag_3w_xf — Function

Calculate dc current on three-winding transformers

DC Current Calculation

PowerModelsGMD.calc_dc_current_line — Function

Calculate dc current for a transmission line based on node voltages from GMD solve

PowerModelsGMD.calc_dc_current_xfmr — Function

Calculate dc current for a transformer based on node voltages from GMD solve

PowerModelsGMD.calc_dc_current_sub — Function

Calculate dc current for a substation based on node voltages from GMD solve

Transformer Temperature

PowerModelsGMD.default_topoil_rated_temp_rise_c — Constant

Transformer top-oil temperature rise in degrees C above ambient at rated power

PowerModelsGMD.default_hotspot_coeff — Constant

Transformer GIC hotspot cofficient in units of degrees C per amp

PowerModelsGMD.default_hotspot_time_const_secs — Constant

Transformer hotspot temperature time constant in seconds

PowerModelsGMD.default_topoil_time_const_mins — Constant

Transformer top-oil temperature time constant in minutes

PowerModelsGMD.default_ambient_temp_c — Constant

Ambient temperature in units of degrees C

PowerModelsGMD.calc_transformer_temps! — Function

Calculate transformer top-toil and hot-spot temperatures from GIC and ac powerflow results

PowerModelsGMD.calc_hotspot_temp! — Function

Calculate absolute hot-spot temperature from top-oil and hot-spot temperature rise

PowerModelsGMD.calc_delta_hotspotrise! — Function

Calculate hotspot temperature rise at the current time step based on GMD solve and hotspot temperature rise at the previous time step

PowerModelsGMD.calc_delta_topoilrise_ss — Function

Calculate steady-state top-oil temperature rise

PowerModelsGMD.calc_delta_topoilrise! — Function

Calculate top-oil temperature rise at the current time step based on ac powerflow solve and hotspot temperature rise at the previous time step

Bounds Calculation

PowerModelsGMD.calc_min_dc_voltage — Function

Calculate the minimum dc voltage at a gmd bus

PowerModelsGMD.calc_max_dc_voltage — Function

Calculate the maximum dc voltage at a gmd bus

PowerModelsGMD.calc_max_q_loss — Function

Calculate the max q loss of xfmr in branch

PowerModelsGMD.calc_ac_positive_current_mag_min — Function

Calculate the minimum absolute value AC current on a branch

PowerModelsGMD.calc_ac_current_mag_max — Function

Calculate the maximum absolute value AC current on a branch

DC Network Building

PowerModelsGMD.generate_dc_data — Function

Load a network dictionary from a set of GIC/RAW/CSV file paths

Load a network dictionary from a pair of GIC/RAW file paths and calculate coupled voltages with a uniform field

Load a network dictionary from a pair of GIC/RAW file handles and calculate coupled voltages with a uniform field

Load a network dictionary from a pair of GIC/RAW dictionary structures and calculate coupled voltages with a uniform field

PowerModelsGMD.generate_dc_data_psse — Function

Load a network dictionary from a pair of GIC/RAW file handles and calculate coupled voltages with a uniform field

PowerModelsGMD.generate_dc_data_matpower — Function

Load a network dictionary from a pair of GIC/MatPower file handles and calculate coupled voltages with a uniform field

PowerModelsGMD.parse_files — Function

Parse a set of files given their paths and return as multinetwork dictionary structure

File type is inferred based on the file extension

PowerModelsGMD.parse_gic — Function

Parse a GIC file given a file path and return as a dictionary structure

Parse a GIC file given a file handle and return as a dictionary structure

PowerModelsGMD.add_coupled_voltages! — Function

Add coupled line voltages from a CSV file path into the branch table

Add coupled line voltages from a file handle into the branch table

Add coupled line voltages from a dataframe into the branch table

Utility Functions

PowerModelsGMD.get_warn — Function

Retrieve value from dictionary and print warning if default is used

PowerModelsGMD.generate_g_i_matrix — Function

Create sparse admittance matrix and current injection vector from network data

PowerModelsGMD.build_adjacency_matrix — Function

Create adjacency matrix from network data